Move non-NLEIGS specific code to rk_helper directory
