+++ src/nleigs/method_nleigs.jl
@@ -1,13 +1,8 @@
+using NonlinearEigenproblems.RKHelper
 using LinearAlgebra
 using Random
 
 export nleigs
-
-include("linsolvercache.jl")
-include("discretizepolygon.jl")
-include("lejabagby.jl")
-include("ratnewtoncoeffs.jl")
-include("ratnewtoncoeffsm.jl")
 
 """
     nleigs(nep::NEP, Σ::AbstractVector{Complex{T}})
